# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_dormitory-life_core_internal_server_request_models.CreateDormitoryEventResponse:
    properties:
      description:
        type: string
      dormitory_id:
        type: string
      event_id:
        type: string
      photos:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreatePhotoResponse'
        type: array
      title:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.CreatePhotoResponse:
    properties:
      fileName:
        type: string
      filePath:
        type: string
      size:
        format: int64
        type: integer
      url:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.DeleteDormitoryPhotosResponse:
    properties:
      dormitoryId:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.ErrorResponse:
    properties:
      details:
        items:
          type: string
        type: array
      error:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.Event:
    properties:
      created_at:
        type: string
      description:
        type: string
      dormitory_id:
        type: string
      event_id:
        type: string
      event_photos:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.FileInfo'
        type: array
      title:
        type: string
    type: object
  github_com_dormitory-life_core_internal_server_request_models.FileInfo:
    properties:
      last_modified:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      url:
        type: string
    type: object
...
  github_com_dormitory-life_core_internal_server_request_models.GetDormitoryEventsResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.Event'
        type: array
    type: object
...
info:
  contact: {}
  description: API Core сервиса Dormitory Life
  title: Dormitory Life Core API
  version: "1.0"
paths:
...
  /core/dormitories/{dormitory_id}/events:
    get:
      description: Получение списка событий общежития
      produces:
      - application/json
      responses:
        "200":
          description: Лента
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.GetDormitoryEventsResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      summary: Получение ленты
      tags:
      - Feed
    post:
      consumes:
      - multipart/form-data
      description: Создает новое событие для общежития с заголовком, описанием и фотографиями
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      - description: Заголовок события
        in: formData
        name: title
        required: true
        type: string
      - description: Описание события
        in: formData
        name: description
        required: true
        type: string
      - description: Фотографии события
        in: formData
        name: photos
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Событие создано
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateDormitoryEventResponse'
        "400":
          description: Некорректные данные формы или отсутствуют обязательные поля
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать событие общежития
      tags:
      - Feed
  /core/dormitories/{dormitory_id}/events/{event_id}:
    delete:
      description: Удаляет событие из ленты общежития
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить событие общежития
      tags:
      - Feed
...
  /core/dormitories/{dormitory_id}/photos:
    delete:
      description: Удаляет фотографии общежития
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      responses:
        "200":
          description: Фото удалены
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.DeleteDormitoryPhotosResponse'
        "400":
          description: Неверные данные / параметры запроса
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить фотографии общежития
      tags:
      - Dormitories
    post:
      consumes:
      - multipart/form-data
      description: Создает новое фото для общежития
      parameters:
      - description: ID общежития
        in: path
        name: dormitory_id
        required: true
        type: string
      - collectionFormat: multi
        description: Фотографии общежития
        in: formData
        items:
          type: file
        name: photos
        required: true
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Событие создано
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.CreateDormitoryEventResponse'
        "400":
          description: Некорректные данные формы или отсутствуют обязательные поля
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "403":
          description: Нет прав на действие
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/github_com_dormitory-life_core_internal_server_request_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать фото общежития
      tags:
      - Dormitories
...
schemes:
- http
- https
swagger: "2.0"